share midicv logic
