+++ gateware/src/top/dsp/top.py
@@ -465,86 +465,6 @@ class TouchMixTop(wiring.Component):
 
         return m
 
-class MidiCVTop(wiring.Component):
-
-    """
-    Simple monophonic MIDI to CV conversion.
-
-    in: (TRS MIDI)
-    out0: Gate
-    out1: V/oct CV
-    out2: Velocity
-    out3: Mod Wheel (CC1)
-    """
-
-    i: In(stream.Signature(data.ArrayLayout(ASQ, 4)))
-    o: Out(stream.Signature(data.ArrayLayout(ASQ, 4)))
-
-    # Note: MIDI is valid at a much lower rate than audio streams
-    i_midi: In(stream.Signature(midi.MidiMessage))
-
-    def elaborate(self, platform):
-        m = Module()
-
-        m.d.comb += [
-            # Always forward our audio payload
-            self.i.ready.eq(1),
-            self.o.valid.eq(1),
-
-            # Always ready for MIDI messages
-            self.i_midi.ready.eq(1),
-        ]
-
-        # Create a LUT from midi note to voltage (output ASQ).
-        lut = []
-        for i in range(128):
-            volts_per_note = 1.0/12.0
-            volts = i*volts_per_note - 5
-            # convert volts to audio sample
-            x = volts/(2**15/4000)
-            lut.append(fixed.Const(x, shape=ASQ)._value)
-
-        # Store it in a memory where the address is the midi note,
-        # and the data coming out is directly routed to V/Oct out.
-        m.submodules.mem = mem = Memory(
-            width=ASQ.as_shape().width, depth=len(lut), init=lut)
-        rport = mem.read_port(transparent=True)
-        m.d.comb += [
-            rport.en.eq(1),
-        ]
-
-        # Route memory straight out to our note payload.
-        m.d.sync += self.o.payload[1].as_value().eq(rport.data),
-
-        with m.If(self.i_midi.valid):
-            msg = self.i_midi.payload
-            with m.Switch(msg.midi_type):
-                with m.Case(midi.MessageType.NOTE_ON):
-                    m.d.sync += [
-                        # Gate output on
-                        self.o.payload[0].eq(fixed.Const(0.5, shape=ASQ)),
-                        # Set velocity output
-                        self.o.payload[2].as_value().eq(
-                            msg.midi_payload.note_on.velocity << 8),
-                        # Set note index in LUT
-                        rport.addr.eq(msg.midi_payload.note_on.note),
-                    ]
-                with m.Case(midi.MessageType.NOTE_OFF):
-                    # Zero gate and velocity on NOTE_OFF
-                    m.d.sync += [
-                        self.o.payload[0].eq(0),
-                        self.o.payload[2].eq(0),
-                    ]
-                with m.Case(midi.MessageType.CONTROL_CHANGE):
-                    # mod wheel is CC 1
-                    with m.If(msg.midi_payload.control_change.controller_number == 1):
-                        m.d.sync += [
-                            self.o.payload[3].as_value().eq(
-                                msg.midi_payload.control_change.data << 8),
-                        ]
-
-        return m
-
 class PSRAMPingPongDelay(wiring.Component):
 
     """
@@ -940,7 +860,7 @@ CORES = {
     "matrix":         (False, Matrix),
     "touchmix":       (True,  TouchMixTop),
     "waveshaper":     (False, DualWaveshaper),
-    "midicv":         (False, MidiCVTop),
+    "midicv":         (False, midi.MonoMidiCV),
     "psram_pingpong": (False, PSRAMPingPongDelay),
     "sram_pingpong":  (False, SRAMPingPongDelay),
     "psram_diffuser": (False, PSRAMDiffuser),
